Refactor teams page, make server component, useUser in teamsView
