+++ src/app/home/teams/page.tsx
@@ -1,16 +1,12 @@
-'use client';
-
-import { useUser } from "@/app/Providers";
 import { Header } from "@/components/Header";
 import { TeamsView } from "./teamsView";
 
 
 export default function Page() {
-  const { id } = useUser();
   return (
     <>
       <Header text="Teams" />
-      <TeamsView userId={id} />
+      <TeamsView />
     </>
   )
 }
